Add - 정보고시 Form 재활용을 위한 컴포넌트 분리 일부 및 디자인 수정
