+++ src/Components/ProductIntroNotice/NoticeForm/NoticeForm.module.css
@@ -1,30 +1,10 @@
-@value sirDPurple, sirGray,sirLGray,sirBorderRadiusM ,sirDgray, sirTableBorderColor, m from "styles/theme.css";
-
-.grid {
-  border: 1px solid sirTableBorderColor;
-}
-
-.gridTitle {
-  font-size: m;
-  padding: 10px 0;
-  border-bottom: 1px solid sirTableBorderColor;
-  background: #fff;
-}
-
-.gridInnerColor {
-  padding: 20px;
-  background: sirLGray;
-}
+@value sirBorderRadiusM, sirDPurple, sirTableBorderColor from "styles/theme.css";
 
 .gridInner {
   padding: 14px 20px;
   background: #fff;
 }
 
-.box {
-  width: 100%;
-  margin-bottom: 10px;
-}
 .rightBox {
   display: flex;
   justify-content: flex-end;
@@ -42,6 +22,13 @@
   cursor: pointer;
 }
 
+.inputTitle {
+  display: flex;
+  align-items: center;
+  color: sirDgray;
+  font-weight: 600;
+}
+
 .input {
   width: 100%;
   margin-bottom: 7px;
@@ -51,26 +38,9 @@
   cursor: pointer;
 }
 
-.grids {
-  margin-right: 10px;
-}
-
-.inputTitle {
-  display: flex;
-  align-items: center;
-  color: sirDgray;
-  font-weight: 600;
-}
-
-.addBtn {
+.box {
   width: 100%;
-  padding: 10px 0;
-  border: 2px solid sirDPurple;
-  font-size: 18px;
-  font-weight: 700;
-  color: sirDPurple;
-  cursor: pointer;
-  margin-top: 20px;
+  margin-bottom: 10px;
 }
 
 .plusBtn {
