+++ src/Components/ProductIntroNotice/ProductIntroNotice.module.scss
@@ -1,0 +1,16 @@
+@value sirDPurple, sirLGray, sirTableBorderColor from "styles/theme.css";
+
+.gridInnerColor {
+  background: sirLGray;
+  row-gap: 20px;
+}
+
+.addBtn {
+  width: 100%;
+  padding: 10px 0;
+  border: 2px solid sirDPurple;
+  font-size: 18px;
+  font-weight: 700;
+  color: sirDPurple;
+  cursor: pointer;
+}
